+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+import addContacts from "./addContacts";
+import getContactById from "./getContactById";
+import listContact from "./listContact";
+import removeContact from "./removeContact";
+import updateContact from "./updateContact";
+
+export default{
+  addContacts,
+  getContactById,
+  listContact,
+  removeContact,
+  updateContact,
+};
