+++ model/contacts/addContacts.js
@@ -1,7 +1,7 @@
 import fs from "fs/promises";
 import path from "path";
 import { randomUUID } from "crypto";
-import contacts from "../contacts.json";
+import contacts from "../../db/contacts.json";
 import { fileURLToPath } from "url";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -10,7 +10,7 @@ const addContact = async ({ name, email, phone }) => {
   const newContact = { id: randomUUID(), name, email, phone };
   contacts.push(newContact);
   await fs.writeFile(
-    path.join(__dirname, "../contacts.json"),
+    path.join(__dirname, "../../db/contacts.json"),
     JSON.stringify(contacts, null, 2)
   );
   return newContact;
