+++ model/contacts/getById.js
@@ -1,4 +1,4 @@
-import contacts from "../contacts.json";
+import contacts from "../../db/contacts.json";
 
 const getContactById = async (contactId) => {
   const [contact] = contacts.filter((contact) => contact.id === contactId);
