+++ model/contacts/removeContacts.js
@@ -1,22 +1,21 @@
 import fs from "fs/promises";
 import path from "path";
-import contacts from "../contacts.json";
+import contacts from "../../db/contacts.json";
 import { fileURLToPath } from "url";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-const updateContact = async (contactId, body) => {
+const removeContact = async (contactId) => {
   const index = contacts.findIndex((contact) => contact.id === contactId);
   if (index !== -1) {
-    const updatedContact = { id: contactId, ...contacts[index], ...body };
-    contacts[index] = updatedContact;
+    const [result] = contacts.splice(index, 1);
     await fs.writeFile(
-      path.join(__dirname, "../contacts.json"),
+      path.join(__dirname, "../../db/contacts.json"),
       JSON.stringify(contacts, null, 2)
     );
-    return updatedContact;
+    return result;
   }
   return null;
 };
 
-export default updateContact;
+export default removeContact;
