+++ model/contacts/updateContacts.js
@@ -1,21 +1,22 @@
 import fs from "fs/promises";
 import path from "path";
-import contacts from "../contacts.json";
+import contacts from "../../db/contacts.json";
 import { fileURLToPath } from "url";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-const removeContact = async (contactId) => {
+const updateContact = async (contactId, body) => {
   const index = contacts.findIndex((contact) => contact.id === contactId);
   if (index !== -1) {
-    const [result] = contacts.splice(index, 1);
+    const updatedContact = { id: contactId, ...contacts[index], ...body };
+    contacts[index] = updatedContact;
     await fs.writeFile(
-      path.join(__dirname, "../contacts.json"),
+      path.join(__dirname, "../../db/contacts.json"),
       JSON.stringify(contacts, null, 2)
     );
-    return result;
+    return updatedContact;
   }
   return null;
 };
 
-export default removeContact;
+export default updateContact;
