+++ routes/api/contacts.js
@@ -1,44 +1,44 @@
-import { Router } from "express";
-import model from "../../model/controllers/operations";
-import { validateCreate, validateUpdate, validateId } from "./validation";
+// import { Router } from "express";
+// import model from "../../model/controllers/operations";
+// import { validateCreate, validateUpdate, validateId } from "../../middleware/validation/validation";
 
-const router = new Router();
+// const router = new Router();
 
-router.get("/", async (req, res, next) => {
-  const contacts = await model.listContacts();
-  res.status(200).json(contacts);
-});
+// router.get("/", async (req, res, next) => {
+//   const contacts = await model.listContacts();
+//   res.status(200).json(contacts);
+// });
 
-router.get("/:id", validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.getContactById(id);
-  if (contact) {
-    return res.status(200).json(contact);
-  }
-  res.status(404).json({ message: "Not Found" });
-});
+// router.get("/:id", validateId, async (req, res, next) => {
+//   const { id } = req.params;
+//   const contact = await model.getContactById(id);
+//   if (contact) {
+//     return res.status(200).json(contact);
+//   }
+//   res.status(404).json({ message: "Not Found" });
+// });
 
-router.post("/", validateCreate, async (req, res, next) => {
-  const newContact = await model.addContact(req.body);
-  res.status(201).json(newContact);
-});
+// router.post("/", validateCreate, async (req, res, next) => {
+//   const newContact = await model.addContact(req.body);
+//   res.status(201).json(newContact);
+// });
 
-router.delete("/:id", validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.removeContact(id);
-  if (contact) {
-    return res.status(200).json({ message: "Contact deleted" });
-  }
-  res.status(404).json({ message: "Not Found" });
-});
+// router.delete("/:id", validateId, async (req, res, next) => {
+//   const { id } = req.params;
+//   const contact = await model.removeContact(id);
+//   if (contact) {
+//     return res.status(200).json({ message: "Contact deleted" });
+//   }
+//   res.status(404).json({ message: "Not Found" });
+// });
 
-router.put("/:id", validateUpdate, validateId, async (req, res, next) => {
-  const { id } = req.params;
-  const contact = await model.updateContact(id, req.body);
-  if (contact) {
-    return res.status(200).json(contact);
-  }
-  res.status(404).json({ message: "Not Found" });
-});
+// router.put("/:id", validateUpdate, validateId, async (req, res, next) => {
+//   const { id } = req.params;
+//   const contact = await model.updateContact(id, req.body);
+//   if (contact) {
+//     return res.status(200).json(contact);
+//   }
+//   res.status(404).json({ message: "Not Found" });
+// });
 
-export default router;
+// export default router;
